Extracted context with cancel to rootContext and rootCancel in root.go
